chore: remove obsolete embeddings stub
