+++ src/store/index.js
@@ -1,4 +1,4 @@
 export * from './config.js';
 export * from './database.js';
-export * from './embeddings.js';
+export * from './embeddings/index.js';
 export * from './items.js';
